% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{taxNet}
\alias{taxNet}
\title{Calculate the tax and net salary by providing anual gross salary}
\usage{
taxNet(
  salary,
  bin = c(0, 36, 144, 300, 420, 660, 960) * 10^3,
  rate = c(0.03, 0.1, 0.2, 0.25, 0.3, 0.35, 0.45)
)
}
\arguments{
\item{salary}{annual gross salary in RMB.}

\item{bin}{the bins for taxing.
The default is c(0, 36,000, 144,000, 300,000, 420,000, 660,000, 960,000).}

\item{rate}{the rate for each taxing bin.
The default is c(0.03, 0.1, 0.2, 0.25, 0.3, 0.35, 0.45).}
}
\value{
a list, including annual gross salary, the tax you should pay, and the net anual salary.
And make a plot, projecting the net salary on the curve.
}
\description{
Calculate the tax and net salary by providing anual gross salary
}
\examples{
taxNet(100000)
taxNet(c(100000, 200000, 250000, 270000, 300000))
}
