% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustax.R
\name{ustax}
\alias{ustax}
\title{How much you can make in a US company}
\usage{
ustax(
  salary = 120000,
  months = 12,
  bonus = salary * 0.08,
  stateTaxRate = 0.05,
  married = T,
  joint = T,
  rate401k = 0.06,
  bonus401k = 0.02,
  insurance = 558,
  dental = 46.26,
  D2Y = 6.36
)
}
\arguments{
\item{salary}{basic salary within certain months, default is the annual salary (120K)}

\item{months}{number of months corresponding to the salary}

\item{bonus}{bonus of the year}

\item{stateTaxRate}{state tax rate}

\item{married}{TRUE if you are married}

\item{joint}{TRUE if you pay for the tax jointly with your spouse}

\item{rate401k}{what is the percentage you save the pre-tax salary in your 401k account}

\item{bonus401k}{what is the extra percentage your company give (except the $1-to-$1 match)
in your 401k account}

\item{dental}{how much you pay for your dental insuarance monthly}

\item{D2Y}{the exchange rate of Dollar to Yuan.}

\item{medical}{how much you pay for your healthy insurance monthly}
}
\description{
How much you can make in a US company
}
\examples{
ustax()
ustax(118000)
ustax(60000, months = 6, joint = T)
ustax(50000, months = 5, joint = F)
ustax(40000, months = 4, joint = F)
ustax(30000, months = 3, joint = F)
ustax(20000, months = 2, joint = F)
ustax(10000, months = 1, joint = F)

}
