% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.R
\name{tax}
\alias{tax}
\title{Calculate the tax by providing annual gross salary
#' @param salary annual salary in RMB.}
\usage{
tax(
  salary,
  bin = rev(c(0, 36, 144, 300, 420, 660, 960)) * 10^3,
  rate = rev(c(0.03, 0.1, 0.2, 0.25, 0.3, 0.35, 0.45))
)
}
\arguments{
\item{bin}{the rate for each taxing bin.}
}
\value{
The tax you should pay.
}
\description{
Calculate the tax by providing annual gross salary
#' @param salary annual salary in RMB.
}
\examples{
tax(100000)
tax(c(100000, 200000, 250000, 270000, 300000))
}
